pyproject.toml
[tool.poetry]
name = "pdf-ingestion-qa"
version = "0.1.0"
description = ""
authors = ["muhammad-stack <142173430+muhammad-stack@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-core = "^0.3.12"
langchain-community = "^0.3.3"
langchain-google-genai = "^2.0.1"
python-dotenv = "^1.0.1"
pypdf = "^5.0.1"
faiss-cpu = "^1.9.0"
streamlit = "^1.39.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
